+++ homeassistant/components/speedtestdotnet/const.py
@@ -1,0 +1,10 @@
+"""Consts used by Speedtest.net."""
+
+DOMAIN = 'speedtestdotnet'
+DATA_UPDATED = '{}_data_updated'.format(DOMAIN)
+
+SENSOR_TYPES = {
+    'ping': ['Ping', 'ms'],
+    'download': ['Download', 'Mbit/s'],
+    'upload': ['Upload', 'Mbit/s'],
+}
