+++ requirements_all.txt
@@ -1566,7 +1566,7 @@ solaredge==0.0.2
 # homeassistant.components.climate.honeywell
 somecomfort==0.5.2
 
-# homeassistant.components.sensor.speedtest
+# homeassistant.components.speedtestdotnet
 speedtest-cli==2.0.2
 
 # homeassistant.components.spider
